@@ -3,14 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@material-ui/core": "^4.6.1",
+    "@material-ui/icons": "^4.5.1",
     "axios": "^0.18.0",
+    "bootstrap": "^4.3.1",
     "classnames": "^2.2.6",
     "font-awesome": "^4.7.0",
     "polished": "^1.9.3",
     "prop-types": "^15.6.2",
-    "react": "^16.4.2",
+    "react": "16.9.0",
+    "react-dom": "16.9.0",
+    "react-alert": "^6.0.0",
+    "react-alert-template-basic": "^1.0.0",
+    "react-bootstrap": "^1.0.0-beta.14",
     "react-dimensions": "^1.3.1",
-    "react-dom": "^16.4.2",
     "react-dropzone": "^5.0.1",
     "react-map-gl": "^3.3.4",
     "react-router-dom": "^4.3.1",
