@@ -1,64 +1,147 @@
-import React, { Component } from "react";
-import { Link, withRouter } from "react-router-dom";
-
+import React from "react";
+import { Link, withRouter, Redirect } from "react-router-dom";
 import Logo from "../../assets/airbnb-logo.svg";
-
 import api from "../../services/api";
+import { login } from "../../services/auth";
+import { Button, TextField } from '@material-ui/core';
+
+import { Alert } from '../../components/alert'
 
 import { Form, Container } from "./styles";
 
-class SignUp extends Component {
-  state = {
-    username: "",
-    email: "",
-    password: "",
-    error: ""
-  };
 
-  handleSignUp = async e => {
-    e.preventDefault();
-    const { username, email, password } = this.state;
-    if (!username || !email || !password) {
-      this.setState({ error: "Preencha todos os dados para se cadastrar" });
-    } else {
-      try {
-        await api.post("/users", { username, email, password });
-        this.props.history.push("/");
-      } catch (err) {
-        console.log(err);
-        this.setState({ error: "Ocorreu um erro ao registrar sua conta. T.T" });
-      }
-    }
-  };
+function SignUp() {
 
-  render() {
+    const [state, setState] = React.useState({
+        username: "",
+        email: "",
+        password: "",
+        confirmPassword: "",
+        redirect: false
+    })
+
+    const [alertState, setAlertState] = React.useState({
+        variant: 'error',
+        mensage: "",
+        alertOpen: false
+    })
+
+    const handleSignUp = async () => {
+        const { username, email, password, confirmPassword } = state;
+
+        if (!email || !password || !username || !confirmPassword) {
+            setAlertState({ variant: 'error', mensage: "Preencha todos os dados para continuar!", alertOpen: true });
+            return;
+        }
+
+        if (password !== confirmPassword) {
+            setAlertState({ variant: 'error', mensage: "Senhas não conferem, por favor verifique!", alertOpen: true });
+            return;
+        }
+
+        if (email.indexOf('@') < 0 || email.indexOf('.') < 0) {
+            setAlertState({ variant: 'error', mensage: "Infome um email valido", alertOpen: true });
+            return;
+        }
+
+
+        try {
+            const response = await api.post("/register", { username, email, password });
+            console.log(response.data)
+            if (response.data.username) {
+                setAlertState({ variant: 'success', mensage: "Registro realizado com sucesso", alertOpen: true });
+                setTimeout(() => {
+                    setState({ ...state, redirect: true })
+                }, 3000)
+
+            }
+        } catch (err) {
+            console.error(err)
+
+            setAlertState({
+                variant: 'error',
+                alertOpen: true,
+                mensage:
+                    "Houve um problema com o login, verifique suas credenciais. T.T"
+            });
+        }
+
+    };
+
+    const handleClose = (event, reason) => {
+        if (reason === 'clickaway') {
+            return;
+        }
+
+        setAlertState({ variant: 'error', mensage: "", alertOpen: false });
+    };
+
     return (
-      <Container>
-        <Form onSubmit={this.handleSignUp}>
-          <img src={Logo} alt="Airbnb logo" />
-          {this.state.error && <p>{this.state.error}</p>}
-          <input
-            type="text"
-            placeholder="Nome de usuário"
-            onChange={e => this.setState({ username: e.target.value })}
-          />
-          <input
-            type="email"
-            placeholder="Endereço de e-mail"
-            onChange={e => this.setState({ email: e.target.value })}
-          />
-          <input
-            type="password"
-            placeholder="Senha"
-            onChange={e => this.setState({ password: e.target.value })}
-          />
-          <button type="submit">Cadastrar grátis</button>
-          <hr />
-          <Link to="/">Fazer login</Link>
-        </Form>
-      </Container>
+        <Container>
+            {state.redirect ? <Redirect to='/' /> : null}
+
+            <Alert
+                open={alertState.alertOpen}
+                onClose={handleClose}
+                variant={alertState.variant}
+                message={alertState.mensage}
+            />
+            <Form>
+                <img src={Logo} alt="Airbnb logo" />
+                <TextField
+                    margin="normal"
+                    required
+                    fullWidth
+                    id="user"
+                    label="Usuario"
+                    name="user"
+                    onChange={(e) => { setState({ ...state, username: e.target.value }) }}
+                    autoFocus
+                />
+                <TextField
+                    margin="normal"
+                    required
+                    fullWidth
+                    id="email"
+                    label="Email"
+                    name="email"
+                    autoComplete="email"
+                    onChange={(e) => { setState({ ...state, email: e.target.value }) }}
+                    autoFocus
+                />
+                <TextField
+                    margin="normal"
+                    required
+                    fullWidth
+                    name="password"
+                    label="Senha"
+                    type="password"
+                    id="password"
+                    autoComplete="current-password"
+                    onChange={(e) => { setState({ ...state, password: e.target.value }) }}
+                />
+
+                <TextField
+                    margin="normal"
+                    required
+                    fullWidth
+                    name="password"
+                    label="Confimação de Senha"
+                    type="password"
+                    id="ConfirmPassword"
+                    onChange={(e) => { setState({ ...state, confirmPassword: e.target.value }) }}
+                />
+
+                <Button variant="contained" color="primary" onClick={handleSignUp}>
+                    Cadastrar
+                 </Button>
+
+                <hr />
+                <Link to="/">Logar</Link>
+            </Form>
+        </Container >
     );
-  }
+
 }
 
 export default withRouter(SignUp);
