@@ -26,27 +26,9 @@ export const Form = styled.form`
     width: 100%;
     text-align: center;
   }
-  input {
-    flex: 1;
-    height: 46px;
-    margin-bottom: 15px;
-    padding: 0 20px;
-    color: #777;
-    font-size: 15px;
-    width: 100%;
-    border: 1px solid #ddd;
-    &::placeholder {
-      color: #999;
-    }
-  }
   button {
-    color: #fff;
-    font-size: 16px;
-    background: #fc6963;
-    height: 56px;
-    border: 0;
-    border-radius: 5px;
     width: 100%;
+    margin-top: 15px;
   }
   hr {
     margin: 20px 0;
